Moved utility functions for testing into /nistats/utils.py from other modules
 - Functions moved from /tests/test_first_level_model.py ino /nistats/utils.py:
   - write_fake_fmri_data()
   - generate_fake_fmri_data()
 - Functions moved from /tests/test_utils.py into /nistats/utils.py:
   - write_fake_bold_img()
   - basic_paradigm()
   - basic_confounds()
   - create_fake_bids_dataset()
